Make persistent browser session runnable type and id nullable add browser_id
