+++ alembic/versions/2025_01_06_1200-2d3b576fb007_make_persistent_browser_session_.py
@@ -1,0 +1,40 @@
+"""make persistent browser session runnable type and id nullable add browser_id
+
+Revision ID: 2d3b576fb007
+Revises: 411dd89f3df9
+Create Date: 2025-01-06 12:00:21.104308+00:00
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '2d3b576fb007'
+down_revision: Union[str, None] = '411dd89f3df9'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('persistent_browser_sessions',
+    sa.Column('persistent_browser_session_id', sa.String(), nullable=False),
+    sa.Column('organization_id', sa.String(), nullable=False),
+    sa.Column('runnable_type', sa.String(), nullable=False),
+    sa.Column('runnable_id', sa.String(), nullable=False),
+    sa.Column('created_at', sa.DateTime(), nullable=False),
+    sa.Column('modified_at', sa.DateTime(), nullable=False),
+    sa.Column('deleted_at', sa.DateTime(), nullable=True),
+    sa.ForeignKeyConstraint(['organization_id'], ['organizations.organization_id'], ),
+    sa.PrimaryKeyConstraint('persistent_browser_session_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('persistent_browser_sessions')
+    # ### end Alembic commands ###
